requirements.txt
streamlit==1.30.0
pandas==2.1.1
requests==2.31.0
pydeck==0.8.0b3
plotly==5.16.1
uvicorn==0.23.2
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
numpy==1.26.0
